scraper, multithreading
# What is new to Snake.

> Snake is basically a web crawler put into a PyQt5 Framework. The software is completely > written by me. Just for fun to be honest with you. Normally people write documentation for their software, so this is my attempt. Just know the software is in early development.

"Snake" has a build in text editor to make life easy. Slide with you buddy "Python3.7" like a snake through the data the internet offers and save it right away to your system. Or as a text editor suggests edit your fetch!

### What is New 

- Scraper advanced search
> The scraper part of snake is now optimized with regex and collect big chunks of data. How bigger the site how bigger the chance on a application freeze. This is something i have to look into for in the future.
- Output results
> Each result has a tree structure. This way you are able to have a quick overview what links are available to you.
- Multithreading
> I'm not sure yet how to tackle the freezing problem, the code behind it keeps running so its not a big problem. Its the GUI that freezes. Perhaps this is my  future solution.

### Fixes
- TimeoutError
> Fixed a timeout error when a invalid link was used.

###### New Updates
> Snake lets you know when there is a new version available to download. Because this is a opensource project, there will be a chance someone uses my code which is fine by me but because this feature is implemented, I want you to know that I (the creator) only upload at:
- https://github.com/Annihilator708/Snake
- https://www.codewars.nl/

> Do not be foolish and download malware.

### Contribution
Ideas and requests are greatly appreciated. If you have any ideas please feel free to use the template "feature_request_template" provided in the **documentation** folder

### Bugs
Like wise Contribution there is a bug template available in the **documentation** folder. Feel free to create a ticket with your problem or contribute to solve the issue. Don't forget to make a pull request!